pyproject.toml

[project]
name = "my_app"
version = "0.1.0"
requires-python = ">=3.12"
description = "REST API for Organization Directory with JWT based authentication using OAuth2 token exchange and Ed25519."
authors = [
  { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
  "fastapi>=0.95.0",
  "uv>=0.5.9",
  "pydantic>=1.9.0",
  "sqlalchemy>=1.4.0",
  "alembic>=1.7.0",
  "python-jose[cryptography]>=3.3.0"
]

[tool.uv]
# uv-specific configuration options, if needed.
